api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.9
pymysql>=1.0.3
email-validator>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.2
google-generativeai>=0.3.0
Pillow>=9.0.0
SpeechRecognition>=3.10.0
mysql-connector-python>=8.0.33